zmiany
dodanie currentUser do podtrzymania dostepu;
dodanie tokenu do wylogowania;
zakomentowanie powielonych zalezności;
zamienione default na null w schemas/user
ujednolicone user-service.js i przeniesione do jednego folderu services
