+++ controllers/loginUser.js
@@ -1,49 +1,45 @@
-const { findUserByEmail, updateToken } = require("../service/user-service")
+const { findUserByEmail, updateToken } = require("../services/user-service");
 const bcrypt = require("bcryptjs");
 const { generateToken } = require("../config/passport-jwt");
 
 const loginUser = async (req, res, next) => {
-    const { password, email} = req.body;
+  const { password, email } = req.body;
 
-    try {
-        const user = await findUserByEmail({email})
+  try {
+    const user = await findUserByEmail({ email });
 
-        if (!user) {
-            await updateToken
-            res.status(401).json({
-              status: "Unauthorized",
-              code: 401,
-              message: `There is no user ${email}`,
-            });
-          }
+    if (!user) {
+      await updateToken;
+      res.status(401).json({
+        status: "Unauthorized",
+        code: 401,
+        message: `There is no user ${email}`,
+      });
+    }
 
-          const isPasswordMatch = await bcrypt.compare(password, user.password);
+    const isPasswordMatch = await bcrypt.compare(password, user.password);
 
-          if (!isPasswordMatch) {
-            return res.status(401).json({
-              status: "Unauthorized",
-              code: 401,
-              message: "Wrong password",
-            });
-          }
+    if (!isPasswordMatch) {
+      return res.status(401).json({
+        status: "Unauthorized",
+        code: 401,
+        message: "Wrong password",
+      });
+    }
 
-          user.token = generateToken(user);
+    user.token = generateToken(user);
     await user.save();
     res.status(200).json({
       message: "Login successful",
       token: user.token,
       user: {
         email: user.email,
-        subscription: user.subscription,
       },
     });
   } catch (e) {
     console.error(e);
     next(e);
-
-    }
-}
-
-
+  }
+};
 
 module.exports = { loginUser };
