--- service/user-service.js
@@ -1,18 +1,0 @@
-const User = require("../schemas/user");
-
-const createUser = ({ email, password, name }) => {
-  return User.create({ email, password, name });
-};
-const findUserByEmail = ({ email }) => {
-  return User.findOne({ email });
-};
-
-const updateToken = ({ email, token }) => {
-  return User.findOneAndUpdate({ email }, { $set: { token } });
-};
-
-module.exports = {
-  updateToken,
-  createUser,
-  findUserByEmail,
-};
